requirements.txt
boto3==1.40.38
botocore==1.40.38
cffi==2.0.0
cryptography==46.0.1
jmespath==1.0.1
MarkupSafe==3.0.2
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
s3transfer==0.14.0
sendgrid==6.12.5
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
